#3 MathLibraryTests
 - fixs from code review

+++ src/Calculator/MathLib/Functions/Advanced/Power.cs
@@ -1,8 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using MathLib.Interfaces;
 
 namespace MathLib.Functions.Advanced
@@ -14,7 +10,7 @@ namespace MathLib.Functions.Advanced
             throw new NotImplementedException();
         }
 
-        public double Calculate(double operand, int exponent)
+        public double Calculate(double operand, double exponent)
         {
             throw new NotImplementedException();
         }
